Split attrs tests into focused modules
